requirements.txt
certifi==2019.11.28
chardet==3.0.4
flatten-json==0.1.7
idna==2.8
kafka-python==1.4.7
PyYAML==5.3
requests==2.22.0
six==1.14.0
urllib3==1.25.7
